requirements.txt
pydantic>=2.0.0
moviepy>=2.0.0
pillow>=10.0.0
pillow-heif>=0.18.0
